Updated README.md with package reference
When compiling your own Docker image you need to pass the correction `RUBY_VERSION` and `RUBY_CHECKSUM`. Previously, the README.md failed to mention which checksum exactly it was the image was using for verifying the source. I've added a small sentence which clarifies that I'm using the `tar.gz` package to build the image.